Begin breaking out a Collection and Properties reusable
